+++ src/adapter/http/src/middleware/authentication_layer.rs
@@ -71,8 +71,8 @@ impl<Svc> AuthenticationMiddleware<Svc> {
                     auth::AccessTokenError::Expired => Ok(CurrentAccountSubject::anonymous(
                         AnonymousAccountReason::AuthenticationExpired,
                     )),
-                    auth::AccessTokenError::Invalid(_) => {
-                        tracing::warn!("Ignoring invalid auth token");
+                    auth::AccessTokenError::Invalid(err) => {
+                        tracing::warn!(error = err, "Ignoring invalid auth token",);
                         Ok(CurrentAccountSubject::anonymous(
                             AnonymousAccountReason::AuthenticationInvalid,
                         ))