+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- `kamu system check-token` command for token debugging
+### Fixed
+- `kamu system api-server` startup failure
+
 ## [0.162.0] - 2024-02-28
 ### Added
 - Flow system now fully supports batching conditions for derived datasets: